Token done

+++ pages/api/products/[id].js
@@ -5,8 +5,9 @@ export default async function handler(req, res) {
   const {
     method,
     query: { id },
+    cookies,
   } = req;
-
+  const token = cookies.token;
   dbConnect();
 
   if (method === "GET") {
@@ -19,6 +20,9 @@ export default async function handler(req, res) {
   }
 
   if (method === "PUT") {
+    if (!token || token !== process.env.TOKEN) {
+      return res.status(401).json("Not authenticated!");
+    }
     try {
       const product = await Product.findByIdAndUpdate(id, req.body, {
         new: true,
@@ -30,6 +34,9 @@ export default async function handler(req, res) {
   }
 
   if (method === "DELETE") {
+    if (!token || token !== process.env.TOKEN) {
+      return res.status(401).json("Not authenticated!");
+    }
     try {
       await Product.findByIdAndDelete(id);
       res.status(200).json("The product has been deleted!");
